% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMalct.R
\name{JMalct}
\alias{JMalct}
\title{Main function to carry out variable allocation for a joint model!}
\usage{
JMalct(y, time, delta, X, t, id, mstop = 500, nu = 0.1)
}
\arguments{
\item{y}{vector containing the longitudinal outcome}

\item{time}{vector containing the observed survival times}

\item{delta}{censoring indicator of the survival times}

\item{X}{design matrix of possibly high dimensional baseline covariates which underly the allocation process}

\item{t}{vector of longitudinal observation times}

\item{id}{cluster variable}

\item{mstop}{number of total boosting iterations, default is mstop = 500}

\item{nu}{learning rate of the single boosting updates, default is nu = 0.1}
}
\description{
Main function to carry out variable allocation for a joint model!
}
